requirements.txt
streamlit>=1.26.0

pandas>=2.0.0
numpy>=1.24.0

flickrapi>=2.4.0
pynytimes>=0.7.0
requests>=2.31.0
requests-oauthlib>=1.3.1

nltk>=3.8.1
contractions>=0.1.73
inflect>=6.0.4
wordcloud>=1.9.2

python-dotenv>=1.0.0

matplotlib>=3.7.2
plotly>=5.17.0
seaborn>=0.12.2

python-dateutil>=2.8.2